Moved coordinate extraction to getObjectPos() function
Added LocationError() exception class.

Added unassigned gesture that activates the relational navigation helper,
that is, plays positional tones for the current mouse cursor position
and  for the navigator object location, so that you can stear the mouse
towards it. When in editable fields, it uses the caret position instead of
the field's own location.

Added more controlling attributes that can later be used for custom settings