New mouse feature: New reference point is available which is the location of a point at which mouse monitor was triggered.
Started on work for integration with easyTableNavigator add-on. Proof of concept (requires modifications to ETN which will be submitted as PR).
installTasks.py already prepared for update with ETN support.

Changes to settings to enable easy manipulation of .show, .save and .skip attributes of Attribute() objects so we may dynamically affect the settings panel and config save or load.

Added SafeFile() object to settings.serialization (for now) that will probably be used in the future to safely and securely save config files.